Migration to VS 2019

+++ app.verify/B.MExistsAccPaths/02.Components.TASGBuilder/TASGBuilder.Fact.Comm.cs
@@ -277,7 +277,7 @@ namespace ExistsAcceptingPath
         if (fromCompStep.kappaStep == meapContext.mu)
         {
           endNodes.Add(fromNode);
-          if (meapContext.MEAPSharedContext.MNP.F.Contains((uint)fromCompStep.qNext))
+          if (meapContext.MEAPSharedContext.MNP.F.Contains(fromCompStep.qNext))
           {
             acceptingNodes.Add(fromNode);
           }
@@ -326,7 +326,7 @@ namespace ExistsAcceptingPath
       {
         ComputationStep compStep = cfg.IdToNodeInfoMap[uNodeId].CompStep;
 
-        if (states.Contains((uint)compStep.qNext))
+        if (states.Contains(compStep.qNext))
         {
           DAGNode v = cfg.NodeEnumeration[uNodeId];
           DAGEdge e = new DAGEdge(edgeId++, v, cfg.t);
